fix: modify Google App Engine credentials
After updating to the last App Engine version the application was
practically useless. It could not be accessed locally nor from the
appspot. Creating a new project but using the same source files and updating
the app id and client id were the only solution that got the app
back up and running. It can now be accessed locally and remotely
just as before the update.