feat: add functionality to query for all sessions and by session size
This is part one of Task 3. We were asked to manually add indexes to
index.yaml that would support the type of queries required by the new
Endpoints methods. Since the earlier tasks required we create sessions
as children of conferences, our app did not generate indexes for
sessions since we would run queries by ancestor. But what if we wanted
to return the set of ALL sessions in datastore? This query could not be
run without first creating the necessary index. An endpoint method was
added as well that would return the set of all sessions independently
of its parent conference. This would be useful for example if we wanted
a listing of all sessions across conferences.

Another index was added to search sessions by maxAttendees. Let's say
a user likes to attend small sessions because she feels she gets more
bang for her buck when it comes to asking questions, since the more
people in a session, the more questions others ask, thus limiting the
number of questions she gets to ask. This index would make it possible
to find all sessions of small size. The accompanying endpoint for
running such a query was also added.