fix: make getFeaturedSpeaker function perform to specs
The specifications originally asked for this function to determine the
featured speaker for a given conference. As it stood, the first
iteration of this function queried all sessions, without regard for
conference to which they belonged. The new function now queries the
conference specified by the websafeConferenceKey and works only on
those sessions belonging to that conference, thus meeting the ori-
ginal specifications.